# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BarberApp API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      parameters:
        - name: shopId
          in: query
          description: Shop owner user ID (required for public booking page)
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /queue:
    get:
      operationId: listQueue
      tags: [queue]
      summary: List queue entries
      responses:
        "200":
          description: Queue list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueEntry"
    post:
      operationId: addToQueue
      tags: [queue]
      summary: Add client to queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueInput"
      responses:
        "201":
          description: Added to queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
  /queue/{id}:
    delete:
      operationId: removeFromQueue
      tags: [queue]
      summary: Remove from queue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /queue/{id}/start:
    post:
      operationId: startQueueEntry
      tags: [queue]
      summary: Start serving a queue entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get barbershop settings
      parameters:
        - name: shopId
          in: query
          description: Shop owner user ID (required for public booking page)
          schema:
            type: string
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update barbershop settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
components:
  schemas:
...
    Service:
      type: object
      required: [id, name, durationMinutes, price, sortOrder, barberIds]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        barberIds:
          type: array
          description: IDs of barbers that perform this service. Empty array means "all barbers" (legacy).
          items:
            type: integer
    ServiceInput:
      type: object
      required: [name, durationMinutes, price]
      properties:
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        price:
          type: number
        imageUrl:
          type: string
        sortOrder:
          type: integer
        barberIds:
          type: array
          items:
            type: integer
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        barberIds:
          type: array
          items:
            type: integer
...
    QueueEntry:
      type: object
      required: [id, clientName, serviceName, servicePrice, serviceDuration, position, status, createdAt]
      properties:
        id:
          type: integer
        appointmentId:
          type: ["integer", "null"]
        clientName:
          type: string
        serviceName:
          type: string
        servicePrice:
          type: number
        serviceDuration:
          type: integer
        position:
          type: integer
        status:
          type: string
        notes:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        createdAt:
          type: string
    QueueInput:
      type: object
      required: [clientName, serviceName, servicePrice, serviceDuration]
      properties:
        clientName:
          type: string
        serviceName:
          type: string
        servicePrice:
          type: number
        serviceDuration:
          type: integer
        notes:
          type: string
...
    DaySchedule:
      type: object
      required: [closed, open, close, lunchStart, lunchEnd]
      properties:
        closed:
          type: boolean
        open:
          type: string
        close:
          type: string
        lunchStart:
          type: string
        lunchEnd:
          type: string
    WeeklySchedule:
      type: object
      required: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
      properties:
        monday:
          $ref: "#/components/schemas/DaySchedule"
        tuesday:
          $ref: "#/components/schemas/DaySchedule"
        wednesday:
          $ref: "#/components/schemas/DaySchedule"
        thursday:
          $ref: "#/components/schemas/DaySchedule"
        friday:
          $ref: "#/components/schemas/DaySchedule"
        saturday:
          $ref: "#/components/schemas/DaySchedule"
        sunday:
          $ref: "#/components/schemas/DaySchedule"
    Settings:
      type: object
      required: [id, barbershopName, ownerName]
      properties:
        id:
          type: integer
        barbershopName:
          type: string
        ownerName:
          type: string
        logoUrl:
          type: ["string", "null"]
          maxLength: 3000000
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        openTime:
          type: ["string", "null"]
        closeTime:
          type: ["string", "null"]
        weeklySchedule:
          oneOf:
            - $ref: "#/components/schemas/WeeklySchedule"
            - type: "null"
        bookingPageMessage:
          type: ["string", "null"]
        paymentEnableNow:
          type: boolean
        paymentEnableOnSite:
          type: boolean
    SettingsUpdate:
      type: object
      properties:
        barbershopName:
          type: string
        ownerName:
          type: string
        logoUrl:
          type: ["string", "null"]
          maxLength: 3000000
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        openTime:
          type: ["string", "null"]
        closeTime:
          type: ["string", "null"]
        weeklySchedule:
          oneOf:
            - $ref: "#/components/schemas/WeeklySchedule"
            - type: "null"
        bookingPageMessage:
          type: ["string", "null"]
        paymentEnableNow:
          type: boolean
        paymentEnableOnSite:
          type: boolean